Align the DB2 client database layer with the pinned blizzard-database-library API (MSVC 2026 build fixes)
The DB2 support code was written against a guessed revision of the
blizzard-database-library and did not compile with the pinned submodule,
surfacing as build errors in ClientDatabase.cpp / ApplicationProject.cpp:

- BlizzardDatabaseTable has Record(unsigned int id) (by id), no RecordById
- GetRecordDefinition() returns std::vector<BlizzardDatabaseRowDefiniton>
  (by value, note the library spelling), not a wrapper with .ColumnDefinitions
  -- iterate the vector directly
- the table has no ColumnCount()/TableRecordDefinition(); drop them together
  with the unused Windows-side getRecordSize()/ColumnCount() declarations
- LoadTable() returns a const& while the table API is not const qualified;
  bridge it with const_cast which compiles against both the pinned and newer
  library revisions
- Build/BlizzardDatabase have no public build id accessor; record the game
  build on NoggitProject (ClientBuildId) when the project database is created
  and return that from buildId() instead of patching library headers
- row definitions carry no isSigned flag; default to signed ints unless the
  dbd type is "uint"
- forward declare BlizzardArchive::ClientData in ApplicationProject.h so the
  header no longer depends on include order
- replace std::format in ClientDatabase.cpp with plain string concatenation
  (MSVC 14.51 C7595, and libstdc++ < 13 has no <format>)
- fix invalid qualified forward declaration 'class Noggit::ClientDatabaseTable;'
  and the unused _row_definition member of a nonexistent type
- make duplicate/unused experimental MapDB.h actually legal C++
- cast qreal to float before std::min/std::max in MapChunk.cpp so MSVC can
  deduce the template argument (issue: QColor::*F() returns double)

Verified with the pinned submodule headers via a type-check harness for every
library interaction we use, plus a full -fsyntax-only pass of
ClientDatabase.cpp with its real project include chain (now zero diagnostics).

+++ src/noggit/MapChunk.cpp
@@ -1401,11 +1401,13 @@ bool MapChunk::stampMCCV(glm::vec3 const& pos, glm::vec4 const& color, float cha
         // transparent mask areas leave the vertex color untouched.
         if (alpha > 0.0f)
         {
+          // note: QColor::*F() return qreal (double), cast before the float
+          // clamps or MSVC cannot deduce the std::min/std::max template type
           glm::vec3 const image_rgb
           {
-            std::min(std::max(image_color.redF() / 0.5f, 0.0f), 2.0f),
-            std::min(std::max(image_color.greenF() / 0.5f, 0.0f), 2.0f),
-            std::min(std::max(image_color.blueF() / 0.5f, 0.0f), 2.0f)
+            std::min(std::max(static_cast<float>(image_color.redF()) / 0.5f, 0.0f), 2.0f),
+            std::min(std::max(static_cast<float>(image_color.greenF()) / 0.5f, 0.0f), 2.0f),
+            std::min(std::max(static_cast<float>(image_color.blueF()) / 0.5f, 0.0f), 2.0f)
           };
 
           mccv[i].x += (image_rgb.x - mccv[i].x) * alpha;
